feat: added custom hook for drawing functionality
